requirements.txt
asgiref==3.8.1
Django==5.1.7
django-debug-toolbar==5.0.1
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.1
mysqlclient==2.2.7
pillow==11.1.0
sqlparse==0.5.3
tzdata==2025.1
